+++ dev/sparktestsupport/modules.py
@@ -482,6 +482,7 @@ pyspark_sql = Module(
         "pyspark.sql.datasource",
         "pyspark.sql.group",
         "pyspark.sql.functions.builtin",
+        "pyspark.sql.functions._hash_funcs",
         "pyspark.sql.functions.partitioning",
         "pyspark.sql.merge",
         "pyspark.sql.readwriter",
@@ -1010,6 +1011,7 @@ pyspark_connect = Module(
         "pyspark.sql.connect.readwriter",
         "pyspark.sql.connect.dataframe",
         "pyspark.sql.connect.functions.builtin",
+        "pyspark.sql.connect.functions._hash_funcs",
         "pyspark.sql.connect.functions.partitioning",
         "pyspark.sql.connect.observation",
         "pyspark.sql.connect.avro.functions",
